tcmu: Add check_config before add device
A patch for tcmu-runner is needed. So add dependency to yasker/tcmu-runner.

Basically tcmu-runner doesn't check if it's the device we want to handle
before open uio device, which caused multiple handlers may contend for the
same uio device(though only one of them should do so).

The patch in tcmu-runner uses check_config to validate the device before
actually open it. Also the handlers in longhorn need to make sure that it only
handle it's own volume, which indicates in dev_config: longhorn//<volume_name>