cli: Only show snapshot of RW replicas
If we merge the result of RW and WO replicas, we will only have the snapshots
after rebuild starting, which is really weird.

During rebuild process, we can only create snapshot. We cannot remove
snapshot or revert to snapshot. So the list of snapshot can be used for backup
purpose only - the backup should goes to RW replica in this case.