frontend: cooperate with longhorn-engine-launcher on snapshot revert and volume creation
Since the launcher is responsible for start/shutdown the frontend (either
block device or iscsi), longhorn-engine need to work with launcher on those.

1. When the frontend is starting for the first time, engine need to notify the
launcher that frontend should be started.

2. When revert a snapshot is requested, engine need to temporaily shutdown the
frontend (blockdev etc) to avoid stale cache. This was observed during CI build.
So now engine will notify launcher to start/shutdown the frontend when revert is
happening.

3. We don't shutdown the frontend of launcher when volume is shuting down,
because it's maybe a live upgrade in process, we don't want to shutdown the
frontend and cause the outage. Launcher will shutdown frontend after launcher
determined that the volume needs to be shutdown.