replica: Don't set index to unknown if the layer was removed
We know what's the index for the newer layers. It has been moved by one after
removing the specified index.

In fact, the context of the layer to be removed SHOULD BE preserved by the newer
layer next to it, but we don't want to make assumption about what caller done.