replica: error out replica if underlying file system doesn't support extent mapping.
Not all file systems support file extent mapping, So replica
should error out if the underlying filesystem does not support
extent mapping.

This commit add check for the extent mapping support.

This can be verified if the given path to the replica is formated
into the file-system which does not give file extent map. Few of
such file-system are FAT, ZFS etc.

Signed-off-by: Utkarsh Mani Tripathi <utkarsh.tripathi@mayadata.io>