sync: revert replica snapshot verification
The snapshot tree of replica can be different during the time of purging
snapshot.

Now we treat the healthy replica with the most snapshots as the source of the
truth, since that means something are still in progress and haven't completed
yet.

The implication is, deletion operation needs to be idempotent now.

Fix https://github.com/rancher/longhorn/issues/354