engine: Add command to export the volume
This commit introduces `longhorn restore-to` command, which gives users the ability to restore a backup to a raw image or a qcow2 image. If the backup is based on a backing file, users should provide the backing file as a qcow2 image with `--backing file` parameter.

Based on https://github.com/rancher/longhorn-engine/pull/325

Signed-off-by: James Oliver <joliver@rancher.com>