Fix creating backup would temporarily change the volume meta head
https://github.com/rancher/longhorn/issues/412

The current code ignored temporarily replica object's `readonly`
properity, result in rewriting the volume metadata file by mistake
while create backup.

But the volume would normally restored to correct situation because
replica.Close() call would write the correct replica info in memory to
the disk. That's why normally this issue won't be exposed. But in the
case of one replica wasn't closed properly, the corrupted volume
metadata would be shown upon restart, result in snapshot failure etc in
the future.

The step to recover the data would be provided in the issue above.