Asynchronous Backup Creation
We are adding a new backup command "backup status <backupID>" to check the backup
status/progress. The backupID argument for this command is the result of
"backup create" command which is used to initiate the backup process.
The output of this command will be in json format with following fields if the values
are non-null: progress, backupURL, backupError