Sync APIs for initiating restore and checking status
The following changes include creating a RestoreStatus structure to store
the restore progress information. Further, this also includes the APIs to
initiate the restoration process and getting the restoration status. These
api's will eventually call the GRPC client and server to trigger the restore
or to get back the result