Revert Replica after Restore Backup
Earlier, the trigger to revert the replica was on the client side as it
was a synchronous operation. Now that the restore is asynchronous, it is
triggered on the server side after the restore is complete. Further, we
are also adding a restriction to perform restore backup only when the
frontend of the volume is disabled.