Use the error msg in backup status to decide if removing the backup status is needed
The previous implementation wrongly checks the error returned by
the func `getBackupStatus()`. This error won't contain substring
`backup not found`, and it may be connection failure etc. We cannot
do anything except for erroring out for this invocation error.

The actual error message we need to check is the error message in
the backup status. It may contain the substring `backup not found`
hence we can use this message to decide if we need to remove the
related backup status.

Longhorn #774

Signed-off-by: Shuo Wu <shuo@rancher.com>