Add test for backup volume list
The test makes sure that if there is failure when trying to retrieve the
backup names for that volume, that the list operation does not fail.
Since a failure would block the list operation of returning information
for all the other backup volumes. Instead of failing the whole operation
we add a Error message to the backup volume.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>