Updated MultiWriterAt.WriteAt to not return 0 on partial success.
MultiWriterAt.WriteAt was returning 0 when one of the WriteAts was
successful.  This did not make sense, so it now returns len(p)
with a set of errors if one of the WriteAts to a replica was
successful and one or more failed.  This is similar to
pkg/controller/replicator.go's ReadAt does.  It tries to find a replica
whose ReadAt succeeds, but it returns the int returned by the successful
ReadAt and a list of errors from the replicas with unsuccessful ReadAts.

This change does not have an impact to the software.  The int value
returned by MultiWriterAt.WriteAt isn't used.  In pkg/dataconn/server.go:

    handleWrite() calls pushResponse() with the return value of WriteAt
      which previously was 0 if a replica failed but now will be the
      len(msg.Data) from original message.
    pushResponse() alters the original message received for sending back
      to tgtd.  It sees that the original message was TypeWrite so it
      sets the Data field in the message to nil because there should be
      no data in a response to a TypeWrite message.  The integer return
      value could be used if err is io.EOF but that doesn't seem
      possible because if the WriteAt goes through MultiWriterAt.WriteAt()
      which would have put an io.EOF into a MultiWriterError. So err
      could never be io.EOF and count won't be used for a TypeWrite
      message.  Therefore for a TypeWrite message, the
      integer return value of any WriteAt() called as a result of
      receiving that message is never used.

This was done as part of the analysis of Longhorn #1917 but doesn't fix
it.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>