Move ping timeout.
The ping timeout was in dataconn.Client, but the logic is changing
to only timeout based on time between responses to TypeRead or TypeWrite
messages.  I moved the ping timeout here to make dataconn.Client simpler
and to handle the ping timeout in the place that sends the ping message.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>