Moved client error/timeout handling into the loop function
I removed the io timer and replaced it with a simple io deadline.
I also moved the io deadline handling into the loop function. instead
of dealing with it as part of the sent/receive functions. Since this
allows for the removal of the global state instead all deadline handling
is done inside of the loop function.

This way handle request/response only deal with valid request/responses
And we no longer need the global client error var.

Signed-off-by: Joshua Moody <joshua.moody@suse.com>