Lazily initialize the replica clients for replica/sync service
This is to optimize the connection count for the longhorn-manager engine
monitor loop, in a single loop the monitor executes the engine binary 8
times this leads to a total connection count of f() = 8e + 6 * 2r = 20
for a volume with 1 replica.

Since all the calls the monitor makes only require either the replica or
the sync client, we can reduce this further with this optimization to
g() = 8e + 6r = 14 per monitor loop.

The monitor loop executes 60/5 = 12 times per minute, so our total
connections per minute are reduced from 240 to 168. For volume with 3
replicas we end up with 312 connection per minute.

No further optimization on this end is possible, the next required task
is removal of the direct engine binary invocations by the longhorn-manager.
This will reduce the connection counts to h() = 1e + 2*r which as one
can see is the desired behavior.

Longhorn #2818

Signed-off-by: Joshua Moody <joshua.moody@suse.com>