Remove ping timeout
The ping timeout was occurring in scenarios with limited bandwidth and
high network usage.  The limited bandwidth and high network traffic was
delaying the transmission of the ping message and the response, causing
the ping timeout to occur when the replica was in fact operational.

Without the ping timeout, the code can still take advantage of the
default TCP keep alive timeout that golang's networking code sets to 15
seconds.  The TCP keep alive timeout is a socket option whose
functionality is implemented by the Linux kernel.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>