Remove fsfreeze due to race condition
We fallback to the prior behavior of calling system sync, which is more
expensive and does not prevent additional io on the filesystem.

As a result of this change, the volume head is no longer dirty after
a snapshot operation unless there was additional io, so we are again
skipping recurring backups if no new io is present.

Longhorn #3125

Signed-off-by: Joshua Moody <joshua.moody@suse.com>