Close TCP connection to replica when closing replica
The TCP socket to the replica was remaining open when the controller was
closing the replica.  This was apparent in the integration tests, where
all the replicas from the controller are removed and the same replica
processes were reload.   During the integration tests, sometimes there
were ten or more TCP connections from a controller to a replica
processes.  This change closes the TCP socket when the controller
removes a replica.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>