controller: Disable reverting to a removing snapshot
Once a snapshot is marked as Removed, the related file may be
modified (snapshot prune) and there is no guarantee of the data
integrity, hence we should disable the revert for this kind of
snapshot.

When users try to revert to a snapshot that is marked as Removed,
the controller just needs to return an error and there is no need
to mark all replicas as ERR. We cannot rely on replicas checking it
since the controller will mark the replica as ERR once the revert
call failed.

Longhorn 3613, 3748

Signed-off-by: Shuo Wu <shuo.wu@suse.com>