Separate service unavailable and replica error
When an engine had no available backend, the controller emitted a message stating "...no available backend...".
Afterward, the volume controller blindly marked all replicas as failed.

To create a RWX volume, it was first created on node A, and then the sharemanager resource was created
with node A as the owner. The sharemanager controller then created a sharemanager pod, which might be located
on node B. After the pod was created, the sharemanager resource's owner was transferred to node B, causing
the volume to detach from node A and attach to node B.

However, if replicas were deleted because of detachment, the in-progress engine instance's startup would fail
due to the "...no available backend..." error. This caused all replicas to be marked as failed (healthyAt is
empty and failedAt is set), and they could not be salvaged anymore. As a result, the volume became faulty.

To avoid the blindly marking failed operation, we should separate the recoverable and unrecoverable error.
For a recoverable error, for example, the replica instance is deleted, the replica is not really failed,
the replica should not be marked as failed directly.

To prevent blindly marking replicas as failed, recoverable and unrecoverable errors can be separated.
The manifest corruption can be regarded as a unrecoverable error. For a recoverable error, for instance,
if a replica instance is deleted and the controller is disconnected with it, it should not be marked as
failed directly since the replica are somehow stopped rather than corrupted. In this case, avoiding marking
all replicas failed can help the volume not trapped in a faulty state and reattach to a node.

Longhorn 5537

Signed-off-by: Derek Su <derek.su@suse.com>
(cherry picked from commit 735a7f5fddc3fb96db33b7ef2d61866ff9b828a2)