feat: increase the timeout of tgt's timeout_handler
To reduce the chance of hitting the read-only issue, the key is to
properly tune the timeout setting for the tgt’s timeout_handler.
Because the SCSI layer limits the number of outstanding write requests,
this ensures that in-flight requests have sufficient time to complete
rather than being prematurely terminated.

Longhorn 8572

Signed-off-by: Derek Su <derek.su@suse.com>